Harden SwingCapture for first device build
- ImpactDetector: cast the CoreMedia flags constant (imported as Int) to UInt32
  to match CMSampleBufferGetAudioBufferListWithRetainedBlockBuffer (would fail
  compilation otherwise)
- ContentView: replace retroactive URL: Identifiable conformance with a ShareItem
  wrapper for the share sheet

+++ GolfLaunchMonitor/SwingCaptureApp/Sources/ImpactDetector.swift
@@ -35,7 +35,7 @@ final class ImpactDetector {
             bufferListSize: MemoryLayout<AudioBufferList>.size,
             blockBufferAllocator: nil,
             blockBufferMemoryAllocator: nil,
-            flags: kCMSampleBufferFlag_AudioBufferList_Assure16ByteAlignment,
+            flags: UInt32(kCMSampleBufferFlag_AudioBufferList_Assure16ByteAlignment),
             blockBufferOut: &blockBuffer
         )
         guard status == noErr, let data = audioBufferList.mBuffers.mData else { return }
